Moved account manager and chainholder to external package
