+++ modApiServer/src/module-info.java
@@ -33,5 +33,6 @@ module aion.apiserver {
     exports org.aion.api.server.http;
     exports org.aion.api.server.http.nano;
     exports org.aion.api.server.http.undertow;
-    exports org.aion.api.server.account;
+    exports org.aion.api.server.external;
+    exports org.aion.api.server.external.account;
 }
