+++ modApiServer/src/org/aion/api/server/http/RpcServer.java
@@ -5,11 +5,11 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import org.aion.api.server.account.AccountManager;
+import org.aion.api.server.external.account.AccountManager;
 import org.aion.api.server.rpc.RpcProcessor;
 import org.aion.api.server.rpc2.Rpc2Shim;
-import org.aion.api.server.rpc3.AionChainHolder;
-import org.aion.api.server.rpc3.ChainHolder;
+import org.aion.api.server.external.AionChainHolder;
+import org.aion.api.server.external.ChainHolder;
 import org.aion.api.server.rpc3.RPCMethods;
 import org.aion.api.server.rpc3.Web3EntryPoint;
 import org.aion.zero.impl.blockchain.AionImpl;
