+++ modApiServer/src/org/aion/api/server/rpc3/MinerStatisticsCalculator.java
@@ -11,6 +11,7 @@ import java.util.LinkedList;
 import java.util.Map;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.stream.Collectors;
+import org.aion.api.server.external.ChainHolder;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.mcf.blockchain.Block;
