+++ modApiServer/src/org/aion/api/server/rpc3/RPCMethods.java
@@ -6,6 +6,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+import org.aion.api.server.external.ChainHolder;
 import org.aion.base.AionTransaction;
 import org.aion.base.AionTxReceipt;
 import org.aion.base.TxUtil;
