+++ modApiServer/test/org/aion/api/server/rpc3/OpsRPCImplTest.java
@@ -12,6 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Function;
+import org.aion.api.server.external.ChainHolder;
 import org.aion.base.AccountState;
 import org.aion.base.AionTransaction;
 import org.aion.base.AionTxReceipt;
