+++ modApiServer/test/org/aion/api/server/rpc3/PersonalRPCImplTest.java
@@ -21,7 +21,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Function;
-import org.aion.api.server.account.AccountManager;
+import org.aion.api.server.external.account.AccountManager;
+import org.aion.api.server.external.AionChainHolder;
+import org.aion.api.server.external.ChainHolder;
 import org.aion.crypto.ECKey;
 import org.aion.crypto.ECKeyFac;
 import org.aion.log.AionLoggerFactory;
